remove !:

+++ src/app/articles/articles.component.ts
@@ -1,7 +1,4 @@
-import {
-  Component,
-  OnInit,
-} from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { ArticleCacheService } from '../article-cache.service';
 import { Article } from '../model/article';
 
@@ -10,8 +7,8 @@ import { Article } from '../model/article';
   templateUrl: './articles.component.html',
   styleUrls: ['./articles.component.css'],
 })
-export class ArticlesComponent implements OnInit{
-  articles!: Article[];
+export class ArticlesComponent implements OnInit {
+  articles: Article[] | undefined;
 
   articlesFilter?: Article[];
 
@@ -22,20 +19,25 @@ export class ArticlesComponent implements OnInit{
     });
   }
 
-  ngOnInit(): void {
-  }
+  ngOnInit(): void {}
   delete(article: Article) {
-    this.articleService.deleteArticle(article.id).subscribe(() => {
-      this.articles = this.articles.filter((a) => a.id !== article.id);
-    });
+    if (this.articles) {
+      this.articleService.deleteArticle(article.id).subscribe(() => {
+        if (this.articles) {
+          this.articles = this.articles.filter((a) => a.id !== article.id);
+        }
+      });
+    }
   }
 
   public searchArticle(e: Event) {
-    const title = (<HTMLInputElement>e.target).value;
-    this.articlesFilter = this.articles.filter(
-      (a) =>
-        a.title.toLowerCase().includes(title.toLowerCase()) ||
-        a.content.includes(title.toLowerCase())
-    );
+    if (this.articles) {
+      const title = (<HTMLInputElement>e.target).value;
+      this.articlesFilter = this.articles.filter(
+        (a) =>
+          a.title.toLowerCase().includes(title.toLowerCase()) ||
+          a.content.includes(title.toLowerCase())
+      );
+    }
   }
 }
